+++ src/Transform.cpp
@@ -5,58 +5,58 @@ AngleValue::AngleValue() {
    degrees = 0.0;
    radians = 0.0;
 }
-double AngleValue::getDegrees() const { return degrees; }
-double AngleValue::getRadians() const { return radians; }
-void AngleValue::setDegrees(double deg) {
+float AngleValue::getDegrees() const { return degrees; }
+float AngleValue::getRadians() const { return radians; }
+void AngleValue::setDegrees(float deg) {
    degrees = deg;
    radians = convertDeg2Rad(deg);
 }
-void AngleValue::setRadians(double rad) {
+void AngleValue::setRadians(float rad) {
    radians = rad;
    degrees = convertRad2Deg(rad);
 }
-double AngleValue::convertRad2Deg(double rad) {
-   double deg = 0.0;
+float AngleValue::convertRad2Deg(float rad) {
+   float deg = 0.0;
    if (rad != 0.0) {
       deg = rad * 180.0 / PI_F;
    }
    return deg;
 }
-double AngleValue::convertDeg2Rad(double deg) {
-   double rad = 0.0;
+float AngleValue::convertDeg2Rad(float deg) {
+   float rad = 0.0;
    if (deg != 0.0) {
       rad = deg * PI_F / 180.0;
    }
    return rad;
 }
 
-AngleDegreeValue::AngleDegreeValue(double deg) { setDegrees(deg); }
+AngleDegreeValue::AngleDegreeValue(float deg) { setDegrees(deg); }
 
-AngleRadValue::AngleRadValue(double rad) { setRadians(rad); }
+AngleRadValue::AngleRadValue(float rad) { setRadians(rad); }
 
-Matrix4d getScale(double x, double y, double z) {
-   Matrix4d ret = Eigen::Matrix4d::Identity();
+Matrix4f getScale(float x, float y, float z) {
+   Matrix4f ret = Eigen::Matrix4f::Identity();
    ret(0, 0) = x;
    ret(1, 1) = y;
    ret(2, 2) = z;
    return ret;
 }
 
-Matrix4d getTranslate(double x, double y, double z) {
-   Matrix4d ret = Eigen::Matrix4d::Identity();
+Matrix4f getTranslate(float x, float y, float z) {
+   Matrix4f ret = Eigen::Matrix4f::Identity();
    ret(0, 3) = x;
    ret(1, 3) = y;
    ret(2, 3) = z;
    return ret;
 }
 
-Matrix4d getRotateEuler(const AngleValue& x, const AngleValue& y,
+Matrix4f getRotateEuler(const AngleValue& x, const AngleValue& y,
                         const AngleValue& z, EulerType type) {
-   Matrix4d retVal = Eigen::Matrix4d::Identity();
+   Matrix4f retVal = Eigen::Matrix4f::Identity();
    Eigen::Matrix3d block;
-   auto rotX = Eigen::AngleAxisd(x.getRadians(), Eigen::Vector3d::UnitX());
-   auto rotY = Eigen::AngleAxisd(y.getRadians(), Eigen::Vector3d::UnitY());
-   auto rotZ = Eigen::AngleAxisd(z.getRadians(), Eigen::Vector3d::UnitZ());
+   auto rotX = Eigen::AngleAxisd(x.getRadians(), Eigen::Vector3f::UnitX());
+   auto rotY = Eigen::AngleAxisd(y.getRadians(), Eigen::Vector3f::UnitY());
+   auto rotZ = Eigen::AngleAxisd(z.getRadians(), Eigen::Vector3f::UnitZ());
    if (type == EulerType::EULER_XYZ) {
       block = rotZ * rotY * rotX;
    } else if (type == EulerType::EULER_ZYX) {
@@ -70,18 +70,18 @@ Matrix4d getRotateEuler(const AngleValue& x, const AngleValue& y,
    return retVal;
 }
 
-Matrix4d getLookAt(const Vector3d& lookFrom, const Vector3d& veclookAt,
-                   const Vector3d& up) {
-   Matrix4d translateMat =
+Matrix4f getLookAt(const Vector3f& lookFrom, const Vector3f& veclookAt,
+                   const Vector3f& up) {
+   Matrix4f translateMat =
        getTranslate(-lookFrom[0], -lookFrom[1], -lookFrom[2]);
-   Vector3d realLookAt = veclookAt.normalized();
-   Vector3d right = realLookAt.cross(up).normalized();
-   Vector3d realUp = right.cross(realLookAt).normalized();
-   Matrix4d rotMat;
+   Vector3f realLookAt = veclookAt.normalized();
+   Vector3f right = realLookAt.cross(up).normalized();
+   Vector3f realUp = right.cross(realLookAt).normalized();
+   Matrix4f rotMat;
    rotMat << right[0], right[1], right[2], 0.0f, realUp[0], realUp[1],
        realUp[2], 0.0f, -realLookAt[0], -realLookAt[1], -realLookAt[2], 0.0f,
        0.0f, 0.0f, 0.0f, 1.0f;
-   Matrix4d lookAt = rotMat * translateMat;
+   Matrix4f lookAt = rotMat * translateMat;
 #ifdef DEBUG
    std::cout << "Print LookAt :" << std::endl;
    std::cout << lookAt << std::endl;
@@ -89,17 +89,17 @@ Matrix4d getLookAt(const Vector3d& lookFrom, const Vector3d& veclookAt,
    return lookAt;
 }
 
-Matrix4d getOrthographic(double l, double r, double t, double b, double n, double f) {
+Matrix4f getOrthographic(float l, float r, float t, float b, float n, float f) {
    // represent left,right,top,buttom,near,far
-   Matrix4d trans = Eigen::Matrix4d::Identity();  // translation matrix
+   Matrix4f trans = Eigen::Matrix4f::Identity();  // translation matrix
    trans << 1, 0, 0, -(r + l) / 2, 0, 1, 0, -(t + b) / 2, 0, 0, 1, -(n + f) / 2,
        0, 0, 0, 1;
 
-   Matrix4d scale = Eigen::Matrix4d::Identity();
+   Matrix4f scale = Eigen::Matrix4f::Identity();
    scale << 2 / (r - l), 0, 0, 0, 0, 2 / (t - b), 0, 0, 0, 0, 2 / (n - f), 0, 0,
        0, 0, 1;
 
-   Matrix4d ortho = scale * trans;
+   Matrix4f ortho = scale * trans;
 #ifdef DEBUG
    std::cout << "Print Orthographic :" << std::endl;
    std::cout << ortho << std::endl;
@@ -107,14 +107,14 @@ Matrix4d getOrthographic(double l, double r, double t, double b, double n, doubl
    return ortho;
 }
 
-Matrix4d getPerspective(const AngleValue& xfov, double aspect, double near,
-                        double far) {
-   double left = tan(xfov.getRadians() / 2.f) * near;
-   double right = -left;
-   double buttom = left * 1.f / aspect;
-   double top = -buttom;
-   Matrix4d ortho = getOrthographic(left, right, top, buttom, near, far);
-   Matrix4d perspective = Eigen::Matrix4d::Identity();
+Matrix4f getPerspective(const AngleValue& xfov, float aspect, float near,
+                        float far) {
+   float left = tan(xfov.getRadians() / 2.f) * near;
+   float right = -left;
+   float buttom = left * 1.f / aspect;
+   float top = -buttom;
+   Matrix4f ortho = getOrthographic(left, right, top, buttom, near, far);
+   Matrix4f perspective = Eigen::Matrix4f::Identity();
    perspective << near, 0, 0, 0, 0, near, 0, 0, 0, 0, near + far, -near * far,
        0, 0, 1, 0;
    perspective = ortho * perspective;
@@ -125,8 +125,8 @@ Matrix4d getPerspective(const AngleValue& xfov, double aspect, double near,
    return perspective;
 }
 
-Matrix4d getViewPort(Vector2i resolution) {
-   Matrix4d viewport;
+Matrix4f getViewPort(Vector2i resolution) {
+   Matrix4f viewport;
    viewport << resolution[0] / 2.f, 0, 0, resolution[0] / 2.f, 0,
        resolution[1] / 2.f, 0, resolution[1] / 2.f, 0, 0, 1, 0, 0, 0, 0, 1;
 #ifdef DEBUG
